[api] break the huge souce_controller in sub controllers
  move /source/:project routes to new controller
  move /source/:project/_meta routes to new controller
  rename the actions to something more REST
  adapt tests and make everything green
