+++ src/api/spec/cassettes/SourceController/PUT_update_project_config/1_4_1.yml
@@ -1,0 +1,104 @@
+---
+http_interactions:
+- request:
+    method: put
+    uri: http://backend:5352/source/home:tom/_config?comment=Updated%20by%20test&user=tom
+    body:
+      encoding: ASCII-8BIT
+      string: comment=Updated+by+test
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+      Content-Length:
+      - '23'
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '21'
+    body:
+      encoding: UTF-8
+      string: '<status code="ok" />
+
+'
+    http_version: 
+  recorded_at: Tue, 05 Jun 2018 12:22:51 GMT
+- request:
+    method: put
+    uri: http://backend:5352/source/home:tom/_config?comment=Updated%20by%20test&user=tom
+    body:
+      encoding: ASCII-8BIT
+      string: comment=Updated+by+test
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+      Content-Length:
+      - '23'
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '21'
+    body:
+      encoding: UTF-8
+      string: '<status code="ok" />
+
+'
+    http_version: 
+  recorded_at: Tue, 05 Jun 2018 12:23:38 GMT
+- request:
+    method: get
+    uri: http://backend:5352/source/home:tom/_config
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - text/plain
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '23'
+    body:
+      encoding: UTF-8
+      string: comment=Updated+by+test
+    http_version: 
+  recorded_at: Tue, 05 Jun 2018 12:25:13 GMT
+recorded_with: VCR 4.0.0
