+++ src/api/test/functional/read_permission_test.rb
@@ -485,14 +485,14 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
   def test_alter_source_access_flags
     # Create public project with protected package
     login_adrian
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> </project>'
     assert_response :success
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> <sourceaccess><disable/></sourceaccess> </project>'
     assert_response 403
     assert_xml_tag tag: 'status', attributes: { code: 'change_project_protection_level' }
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:PublicProject'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:PublicProject'),
         params: '<project name="home:adrian:PublicProject"> <title/> <description/> </project>'
     assert_response :success
     put url_for(controller: :source, action: :update_package_meta, project: 'home:adrian:PublicProject', package: 'pack'),
@@ -516,15 +516,15 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
   def test_alter_access_flags
     # Create public project with protected package
     login_adrian
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> </project>'
     assert_response :success
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> <access><disable/></access> </project>'
     assert_response 403
     assert_xml_tag tag: 'status', attributes: { code: 'change_project_protection_level' }
     login_king
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> <access><disable/></access> </project>'
     assert_response :success
     delete '/source/home:adrian:Project'
@@ -534,7 +534,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
   def test_project_links_to_sourceaccess_protected_package
     # Create public project with protected package
     login_adrian
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:PublicProject'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:PublicProject'),
         params: '<project name="home:adrian:PublicProject"> <title/> <description/> </project>'
     assert_response :success
     # rubocop:disable Metrics/LineLength
@@ -552,7 +552,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     assert_response 403
     assert_xml_tag tag: 'status', attributes: { code: 'project_copy_no_permission' }
     # try to access it via own project link
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:tom:temp'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:tom:temp'),
         params: '<project name="home:tom:temp"> <title/> <description/> <link project="home:adrian:PublicProject"/> </project>'
     assert_response :success
     get '/source/home:tom:temp/ProtectedPackage'
@@ -616,7 +616,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     login_adrian
     get '/source/home:adrian:ProtectedProject'
     assert_response 404
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject'),
         params: '<project name="home:adrian:ProtectedProject"> <title/> <description/> <sourceaccess><disable/></sourceaccess>  </project>'
     assert_response :success
     put url_for(controller: :source, action: :update_package_meta, project: 'home:adrian:ProtectedProject', package: 'Package'),
@@ -628,7 +628,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     get '/source/home:adrian:ProtectedProject/Package'
     assert_response 403
     # try to access it via own project link
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:tom:temp'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:tom:temp'),
         params: '<project name="home:tom:temp"> <title/> <description/> <link project="home:adrian:ProtectedProject"/> </project>'
     assert_response :success
     get '/source/home:tom:temp/Package'
@@ -659,53 +659,53 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     login_tom
 
     # try to link to an access protected hidden project from sourceaccess project
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:tom:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:tom:ProtectedProject2'),
         params: '<project name="home:tom:ProtectedProject2"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:tom:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:tom:ProtectedProject2'),
         params: '<project name="home:tom:ProtectedProject2"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
 
     login_adrian
     # try to link to an access protected hidden project from sourceaccess project
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
         params: '<project name="home:adrian:ProtectedProject2"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
         params: '<project name="home:adrian:ProtectedProject2"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
 
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
         params: '<project name="home:adrian:ProtectedProject2"> <title/> <description/> <sourceaccess><disable/></sourceaccess> </project>'
     assert_response :success
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
         params: '<project name="home:adrian:ProtectedProject1"> <title/> <description/> </project>'
     assert_response :success
 
     # Allow linking from not sourceaccess protected project to protected own. src.rpms are not delivered by the backend.
     #
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
         params: '<project name="home:adrian:ProtectedProject1"> <title/> <description/> <link project="home:adrian:ProtectedProject2"/> </project>'
     assert_response :success
 
     # try to link to an access protected hidden project from access hidden project
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject3'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject3'),
         params: '<project name="home:adrian:ProtectedProject3"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
 
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject3'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject3'),
         params: '<project name="home:adrian:ProtectedProject3"> <title/> <description/> <access><disable/></access> </project>'
     assert_response :success
 
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject3'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject3'),
         params: '<project name="home:adrian:ProtectedProject3"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
 
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject4'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject4'),
         params: '<project name="home:adrian:ProtectedProject4"> <title/> <description/> <access><disable/></access> </project>'
     assert_response :success
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject4'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject4'),
         params: '<project name="home:adrian:ProtectedProject4"> <title/> <description/> <access><disable/></access> <link project="home:adrian:ProtectedProject2"/> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response :success
@@ -714,7 +714,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     login_tom
 
     # try to link to an access protected hidden project
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:tom:temp2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:tom:temp2'),
         params: '<project name="home:tom:temp2"> <title/> <description/> <link project="HiddenProject"/> </project>'
     assert_response 404
 
@@ -754,7 +754,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     login_adrian
     get '/source/home:adrian:ProtectedProject'
     assert_response 404
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject'),
         params: '<project name="home:adrian:ProtectedProject"> <title/> <description/> <access><disable/></access> </project>'
     assert_response :success
     get '/source/home:adrian:ProtectedProject'
@@ -789,18 +789,18 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     prepare_request_with_user 'binary_homer', 'buildservice'
 
     # check if sufficiently protected projects can access protected projects
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject1'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject1'),
         params: '<project name="home:binary_homer:ProtectedProject1"> <title/> <description/> <binarydownload><disable/></binarydownload> </project>'
     assert_response 200
 
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject1'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject1'),
         params: '<project name="home:binary_homer:ProtectedProject1"> <title/> <description/> <repository name="BinaryprotectedProjectRepo"> <path repository="nada" project="BinaryprotectedProject"/> <arch>i586</arch> </repository> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response 200
 
     # check if sufficiently protected projects can access protected projects
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject2'),
         params: '<project name="home:binary_homer:ProtectedProject2"> <title/> <description/> </project>'
     assert_response 200
 
@@ -818,7 +818,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
 
     # check if unsufficiently permitted users tries to access protected projects
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:tom:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:tom:ProtectedProject2'),
         params: '<project name="home:tom:ProtectedProject2"> <title/> <description/>  <repository name="HiddenProjectRepo"> <path repository="nada" project="HiddenProject"/> <arch>i586</arch> </repository> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response 404
@@ -826,44 +826,44 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     # try to access it with a user permitted for access
     login_adrian
 
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
         params: '<project name="home:adrian:ProtectedProject1"> <title/> <description/> <access><disable/></access> </project>'
     # STDERR.puts(@response.body)
     assert_response 200
 
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject1'),
         params: '<project name="home:adrian:ProtectedProject1"> <title/> <description/> <repository name="HiddenProjectRepo"> <path repository="nada" project="HiddenProject"/> <arch>i586</arch> </repository> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response 404
 
     # building against
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
         params: '<project name="home:adrian:ProtectedProject2"> <title/> <description/> <repository name="HiddenProjectRepo"> <path repository="nada" project="HiddenProject"/> <arch>i586</arch> </repository> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response 404
 
     # check if download protected project has to access protected project, which reveals Hidden project existence to others and is and error
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
         params: '<project name="home:adrian:ProtectedProject2"> <title/> <description/> <binarydownload><disable/></binarydownload> </project>'
     assert_response 200
 
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:ProtectedProject2'),
         params: '<project name="home:adrian:ProtectedProject2"> <title/> <description/> <repository name="HiddenProjectRepo"> <path repository="nada" project="HiddenProject"/> <arch>i586</arch> </repository> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response 404
 
     # check if access protected project has access binarydownload protected project
     prepare_request_with_user 'binary_homer', 'buildservice'
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject3'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject3'),
         params: '<project name="home:binary_homer:ProtectedProject3"> <title/> <description/> <access><disable/></access> </project>'
     # STDERR.puts(@response.body)
     assert_response 200
 
     # rubocop:disable Metrics/LineLength
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject3'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:binary_homer:ProtectedProject3'),
         params: '<project name="home:binary_homer:ProtectedProject3"> <title/> <description/> <repository name="BinaryprotectedProjectRepo"> <path repository="nada" project="BinaryprotectedProject"/> <arch>i586</arch> </repository> </project>'
     # rubocop:enable Metrics/LineLength
     assert_response 200
@@ -955,7 +955,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
   def test_setup_default_project
     # Create public project
     login_adrian
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> </project>'
     assert_response :success
     get '/source/home:adrian:Project/_meta'
@@ -969,14 +969,14 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     orig_value = c.default_access_disabled
     c.default_access_disabled = true
     c.save!
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> </project>'
     assert_response :success
     get '/source/home:adrian:Project/_meta'
     assert_response :success
     assert_xml_tag(tag: 'disable', parent: { tag: 'access' })
     # enabling access is not allowed in this mode
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> </project>'
     assert_response 403
     assert_xml_tag tag: 'status', attributes: { code: 'change_project_protection_level' }
@@ -984,7 +984,7 @@ class ReadPermissionTest < ActionDispatch::IntegrationTest
     # enabling access is allowed
     c.default_access_disabled = orig_value
     c.save!
-    put url_for(controller: :source, action: :update_project_meta, project: 'home:adrian:Project'),
+    put url_for(controller: :source_project, action: :update_project_meta, project: 'home:adrian:Project'),
         params: '<project name="home:adrian:Project"> <title/> <description/> </project>'
     assert_response :success
